requirements/default.txt
ruamel.yaml >= 0.11.14,  < 0.12.0
docopt      >= 0.6.2,    < 0.7.0
